Move networking classes to com.mapzen.tangram.networking
