+++ platforms/android/tangram/src/main/java/com/mapzen/tangram/networking/Tls12SocketFactory.java
@@ -1,4 +1,4 @@
-package com.mapzen.tangram;
+package com.mapzen.tangram.networking;
 
 import java.io.IOException;
 import java.net.InetAddress;
